Module Eio

Effects based parallel IO for OCaml.

Eio provides support for concurrency (juggling many tasks) and parallelism (using multiple CPU cores for performance).

It provides facilities for creating and coordinating fibers (light-weight threads) and domains (for parallel processing), as well as interfaces for interacting with resources provided by the operating system.

These features must be used within an event loop, provided by an Eio backend. Applications can use Eio_main.run to run a suitable loop.

See https://github.com/ocaml-multicore/eio for a tutorial.

module Std : sig ... end

Commonly used standard features. This module is intended to be opened.

Fibers

module Switch : sig ... end

Grouping fibers and other resources so they can be turned off together.

module Fiber : sig ... end

A fiber is a light-weight thread.

module Cancel : sig ... end

Cancelling fibers.

Concurrency primitives

module Promise : sig ... end

A promise is a placeholder for result that will arrive in the future.

module Semaphore : sig ... end

A counting semaphore.

module Mutex : sig ... end

Mutual exclusion.

module Condition : sig ... end

Waiting for a condition to become true.

module Lazy : sig ... end

Delayed evaluation.

Collections

module Stream : sig ... end

A stream/queue.

module Pool : sig ... end

A pool of resources.

Multiple domains

module Domain_manager : sig ... end

Parallel computation across multiple CPU cores.

module Executor_pool : sig ... end

A pool of domains for executing jobs.

Errors and debugging

val traceln : ?__POS__:(string * int * int * int) -> ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

traceln fmt outputs a debug message (typically to stderr).

Trace messages are printed by default and do not require logging to be configured first. The message is printed with a newline, and is flushed automatically. traceln is intended for quick debugging rather than for production code.

Unlike most Eio operations, traceln will never switch to another fiber; if the OS is not ready to accept the message then the whole domain waits.

It is safe to call traceln from multiple domains at the same time. Each line will be written atomically.

Examples:

traceln "x = %d" x;
traceln "x = %d" x ~__POS__;   (* With location information *)

parameter __POS__
Display __POS__ as the location of the traceln call.

module Exn : sig ... end

Eio exceptions.

exception Io of Exn.err * Exn.context

module Debug : sig ... end

Control over debugging.

Cross-platform OS API

The general pattern here is that each type of resource has a set of functions for using it, plus a provider (Pi) module to allow defining your own implementations.

The system resources are available from the environment argument provided by your event loop (e.g. Eio_main.run).

module Resource : sig ... end

Defines the base resource type.

Byte streams

module Flow : sig ... end

A flow can be used to read or write bytes.

module Buf_read : sig ... end

Buffered input and parsing.

module Buf_write : sig ... end

Buffered output and formatting.

Networking

module Net : sig ... end

Network sockets and addresses.

File-systems

module Path : sig ... end

Accessing paths on a file-system.

module File : sig ... end

Operations on open files.

module Fs : sig ... end

File-system types.

Processes

module Process : sig ... end

Managing child processes.

Time

module Time : sig ... end

Clocks, time, sleeping and timeouts.

Main env

module Stdenv : sig ... end

The standard environment of a process.

Provider API for OS schedulers

module Private : sig ... end

API for use by the scheduler implementation.