Module Eio.Fiber

A fiber is a light-weight thread.

Within a domain, only one fiber can be running at a time. A fiber runs until it performs an IO operation (directly or indirectly). At that point, it may be suspended and the next fiber on the run queue runs.

val both : ( unit -> unit ) -> ( unit -> unit ) -> unit

both f g runs f () and g () concurrently.

They run in a new cancellation sub-context, and if either raises an exception, the other is cancelled. both waits for both functions to finish even if one raises (it will then re-raise the original exception).

f runs immediately, without switching to any other thread. g is inserted at the head of the run-queue, so it runs next even if other threads are already enqueued. You can get other scheduling orders by adding calls to yield in various places. e.g. to append both fibers to the end of the run-queue, yield immediately before calling both.

If both fibers fail, Exn.combine is used to combine the exceptions.

val pair : ( unit -> 'a ) -> ( unit -> 'b ) -> 'a * 'b

pair f g is like both, but returns the two results.

val all : ( unit -> unit ) list -> unit

all fs is like both, but for any number of fibers. all [] returns immediately.

val first : ( unit -> 'a ) -> ( unit -> 'a ) -> 'a

first f g runs f () and g () concurrently.

They run in a new cancellation sub-context, and when one finishes the other is cancelled. If one raises, the other is cancelled and the exception is reported.

As with both, f runs immediately and g is scheduled next, ahead of any other queued work.

If both fibers fail, Exn.combine is used to combine the exceptions.

val any : ( unit -> 'a ) list -> 'a

any fs is like first, but for any number of fibers.

any [] just waits forever (or until cancelled).

val await_cancel : unit -> 'a

await_cancel () waits until cancelled.

raises Cancel.Cancelled

val fork : sw:Switch.t -> ( unit -> unit ) -> unit

fork ~sw fn runs fn () in a new fiber, but does not wait for it to complete.

The new fiber is attached to sw (which can't finish until the fiber ends).

The new fiber inherits sw's cancellation context. If the fiber raises an exception, Switch.fail sw is called. If sw is already off then fn fails immediately, but the calling thread continues.

fn runs immediately, without switching to any other fiber first. The calling fiber is placed at the head of the run queue, ahead of any previous items.

val fork_sub : sw:Switch.t -> on_error:( exn -> unit ) -> ( Switch.t -> unit ) -> unit

fork_sub ~sw ~on_error fn is like fork, but it creates a new sub-switch for the fiber.

This means that you can cancel the child switch without cancelling the parent. This is a convenience function for running Switch.run inside a fork.

parameter on_error
This is called if the fiber raises an exception. If it raises in turn, the parent switch is failed. It is not called if the parent sw itself is cancelled.

val fork_on_accept : on_handler_error:( exn -> unit ) -> sw:Switch.t -> ( Switch.t -> 'a ) -> ( Switch.t -> 'a -> unit ) -> unit

fork_on_accept ~sw accept handle ~on_handler_error creates a new sub-switch t. It runs accept t in the current fiber and, on success, runs handle t result in a new fiber. It is useful for e.g. accepting network connections, where we need to provide a switch for the new client socket before we have forked, but then move it to a child fiber later.

If accept raises an exception then the effect is the same as Switch.run accept. If handle raises an exception, it is passed to on_handler_error. If that raises in turn, the parent switch is failed. on_handler_error is not called if the parent sw is itself cancelled.

val fork_promise : sw:Switch.t -> ( unit -> 'a ) -> 'a Promise.or_exn

fork_promise ~sw fn schedules fn () to run in a new fiber and returns a promise for its result.

This is just a convenience wrapper around fork. If fn raises an exception then the promise is resolved to the error, but sw is not failed.

val check : unit -> unit

check () checks that the fiber's context hasn't been cancelled. Many operations automatically check this before starting.

raises Cancel.Cancelled
if the fiber's context has been cancelled.

val yield : unit -> unit

yield () asks the scheduler to switch to the next runnable task. The current task remains runnable, but goes to the back of the queue. Automatically calls check just before resuming.