Module Eio.Fs

File-system types.

Note: file-system operations, such as opening or deleting files, can be found in the Path module.

Types

type path = string

type error =

| Already_exists of Exn.Backend.t

| Not_found of Exn.Backend.t

| Permission_denied of Exn.Backend.t

| File_too_large

| Not_native of string
(*
Raised by Path.native_exn.
*)

type Exn.err +=

| E of error

val err : error -> exn

type create = [

| `Never
(*
fail if the named file doesn't exist
*)

| `If_missing of File.Unix_perm.t
(*
create if file doesn't already exist
*)

| `Or_truncate of File.Unix_perm.t
(*
any existing file is truncated to zero length
*)

| `Exclusive of File.Unix_perm.t
(*
always create; fail if the file already exists
*)

]

When to create a new file.

If a new file is created, the given permissions are used for it.

type dir_ty = [

| `Dir

]

type 'a dir = [> dir_ty ] as 'a Std.r

Note: use the functions in Path to access directories.

Provider Interface

module Pi : sig ... end