Module Eio.Generic

A base class for objects that can be queried at runtime for extra features.

type 'a ty = ..

An 'a ty is a query for a feature of type 'a.

class type t = object ... end

val probe : t -> 'a ty -> 'a option

probe t feature checks whether t supports feature. This is mostly for internal use. For example, Eio_unix.FD.peek uses this to get the underlying Unix file descriptor from a flow.