Module Eio.Private

API for use by the scheduler implementation.

module Ctf : sig ... end

This library is used to write event traces in mirage-profile's CTF format.

module Fiber_context : sig ... end

Every fiber has an associated context.

module Effect : sig ... end

Temporary hack for compatibility with ocaml.4.12+domains

module Effects : sig ... end

module Suspend : sig ... end

Suspend a fiber and enter the scheduler.

module Waiters : sig ... end

A queue of fibers waiting for an event.

val traceln_mutex : Stdlib.Mutex.t

The mutex used to prevent two domains writing to stderr at once.

This might be useful if you want to write to it directly yourself, e.g. for a log reporter.

val default_traceln : ?__POS__:(string * int * int * int) -> ( 'a, Stdlib.Format.formatter, unit, unit ) Stdlib.format4 -> 'a

default_traceln is a suitable default implementation for Eio.Std.traceln.

It writes output to stderr, prefixing each line with a "+". If __POS__ is given, it also displays the file and line number from that. It uses mutex so that only one domain's output is written at a time.