Module Private.Effects

type 'a enqueue = ('a, exn) Stdlib.result -> unit

A function provided by the scheduler to reschedule a previously-suspended thread.

type Eio__Effect.eff +=

| Suspend : (Fibre_context.t -> 'a enqueue -> unit) -> 'a Stdlib__effectHandlers.eff	(*
Suspend fn is performed when a fibre must be suspended (e.g. because it called Promise.await on an unresolved promise). The effect handler runs fn fibre enqueue in the scheduler context, passing it the suspended fibre's context and a function to resume it. fn should arrange for enqueue to be called once the thread is ready to run again.
*)

| Fork : Fibre_context.t * (unit -> unit) -> unit Stdlib__effectHandlers.eff	(*
See Fibre.fork
*)

| Trace : (?__POS__:(string * int * int * int) -> ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a) Stdlib__effectHandlers.eff	(*
perform Trace fmt writes trace logging to the configured trace output. It must not switch fibres, as tracing must not affect scheduling. If the system is not ready to receive the trace output, the whole domain must block until it is.
*)

| Get_context : Fibre_context.t Stdlib__effectHandlers.eff