Module Private.Fibre_context

Every fibre has an associated context.

type t

val make_root : unit -> t

Make a new root context for a new domain.

val make : cc:Cancel.t -> t

make ~cc is a new fibre context, initially attached to the given cancellation context.

val destroy : t -> unit

destroy t removes t from its cancellation context.

val tid : t -> int

val cancellation_context : t -> Cancel.t

cancellation_context t is t's current cancellation context.

val set_cancel_fn : t -> (exn -> unit) -> unit

set_cancel_fn t fn sets fn as the fibre's cancel function. If the cancellation context is cancelled, the function is removed and called. When the operation completes, you must call clear_cancel_fn to remove it.

val clear_cancel_fn : t -> bool

clear_cancel_fn t removes the function previously set with set_cancel_fn, if any. Returns true if this call removed the function, or false if there wasn't one. This operation is atomic and thread-safe. An operation that completes in another domain must use this to indicate that the operation is finished (can no longer be cancelled) before enqueuing the result. If it returns false, the operation was cancelled first and the canceller has called (or is calling) the function. If it returns true, the caller is responsible for any resources owned by the function, such as the continuation.

val get_error : t -> exn option

get_error t is Cancel.get_error (cancellation_context t)