Module Time.Mono

Monotonic clocks.

Monotonic clocks are unaffected by corrections to the real-time clock, and so are a better choice for timeouts or measuring intervals, where the absolute time doesn't matter.

A monotonic clock may or may not include time while the computer is suspended.

class virtual t : object ... end

val now : t -> Mtime.t

now t is the current time according to t.

val sleep_until : t -> Mtime.t -> unit

sleep_until t time waits until time before returning.

val sleep : t -> float -> unit

sleep t d waits for d seconds.

val sleep_span : t -> Mtime.span -> unit

sleep_span t d waits for duration d.