Module Eio_mock

Mocks for testing.

When testing an Eio program it is often convenient to use mock resources rather than real OS-provided ones. This allows precise control over the test, such as adding delays or simulated faults. You can always just implement the various Eio types directly, but this module provides some convenient pre-built mocks, and some helpers for creating your own mocks.

Mocks typically use Eio.traceln to record how they were used. This output can be recorded and compared against a known-good copy using e.g. ocaml-mdx.

Mocks may require configuration. For example, a source flow needs to know what data to return when the application reads from it. This can be done using the various on_* functions. For example:

let stdin = Eio_mock.Flow.make "stdin" in
let stdout = Eio_mock.Flow.make "stdout" in
Eio_mock.Flow.on_read stdin [
  `Return "chunk1";
  `Return "chunk2";
  `Raise End_of_file
];
Eio.Flow.copy stdin stdout

This will produce:

+stdin: read "chunk1"
+stdout: wrote "chunk1"
+stdin: read "chunk2"
+stdout: wrote "chunk2"

Configuration

module Action : sig ... end

Actions that can be performed by mock handlers.

module Handler : sig ... end

Control how a mock responds.

Pre-defined mocks

module Flow : sig ... end

Mock Eio.Flow sources and sinks.

module Net : sig ... end

Mock Eio.Net networks and sockets.