Module Eio_mock.Flow

Mock Eio.Flow sources and sinks.

type copy_method = [

| `Read_into
(*
Use the source's read_into method (the default).
*)

| `Read_source_buffer
(*
Use the Eio.Flow.Read_source_buffer optimisation.
*)

]

type ty = [ `Generic | `Mock ] Eio.Net.stream_socket_ty

type t = ty Eio.Std.r

val make : ?pp:string Fmt.t -> string -> t

make label is a mock Eio flow. It can be used as a source, sink, or two-way flow.

parameter pp
Printer to use to display the data.

val on_read : t -> string Handler.actions -> unit

on_read t actions configures the values to return from the mock's read function.

val on_copy_bytes : t -> int Handler.actions -> unit

on_copy_bytes t actions configures the number of bytes to copy in each iteration.

val set_copy_method : t -> copy_method -> unit

set_copy_method t m configures t to use the given method to read from a source during a copy operation.