Module Eio_unix

Extension of Eio for integration with OCaml's Unix module.

Note that OCaml's Unix module is not safe, and therefore care must be taken when using these functions. For example, it is possible to leak file descriptors this way, or to use them after they've been closed, allowing one module to corrupt a file belonging to an unrelated module.

val await_readable : Unix.file_descr -> unit

await_readable fd blocks until fd is readable (or has an error).

val await_writable : Unix.file_descr -> unit

await_writable fd blocks until fd is writable (or has an error).

module FD : sig ... end

Get a Unix.file_descr from an Eio object.

module Ipaddr : sig ... end

Convert between Eio.Net.Ipaddr and Unix.inet_addr.

module Private : sig ... end

API for Eio backends only.

module Ctf : sig ... end