Module Eio_unix.FD

Get a Unix.file_descr from an Eio object.

val peek : Eio.Generic.#t -> Unix.file_descr option

peek x is the Unix file descriptor underlying x, if any. The caller must ensure that they do not continue to use the result after x is closed.

val take : Eio.Generic.#t -> Unix.file_descr option

take x is like peek, but also marks x as closed on success (without actually closing the FD). x can no longer be used after this, and the caller is responsible for closing the FD.