Module Eio_unix.File

Extended file API with support for file descriptors.

type rw_ty = [

| `Unix_fd

| Eio.File.rw_ty

]

type ro_ty = [

| `Unix_fd

| Eio.File.ro_ty

]

val import_rw : sw:Eio.Std.Switch.t -> close_unix:bool -> Unix.file_descr -> [< rw_ty ] Eio.Std.r

import_rw ~sw ~close_unix fd is a read/write Eio file that uses fd.

The file resource will be closed when sw finishes.

The backend takes ownership of fd and may change whether it is non-blocking.

The close_unix and sw arguments are passed to Fd.of_unix.

val import_ro : sw:Eio.Std.Switch.t -> close_unix:bool -> Unix.file_descr -> [< ro_ty ] Eio.Std.r

import_ro is like import_rw, but casts the result to be read-only.