Module Eio_unix.Stdenv

The set of resources provided to a process on a Unix-compatible system.

type base = < stdin : source ; stdout : sink ; stderr : sink ; net : Eio.Net.t ; domain_mgr : Eio.Domain_manager.t ; process_mgr : Process.mgr ; clock : Eio.Time.clock ; mono_clock : Eio.Time.Mono.t ; fs : Eio.Fs.dir Eio.Path.t ; cwd : Eio.Fs.dir Eio.Path.t ; secure_random : Eio.Flow.source ; debug : Eio.Debug.t ; backend_id : string >

The common set of features provided by all traditional operating systems (BSDs, Linux, Mac, Windows).

You can use the functions in Eio.Stdenv to access these fields if you prefer.