Module Eio.Net

Networking.

Example:

let addr = `Tcp (Ipaddr.V4.loopback, 8080)

let http_get ~net ~stdout addr =
  Switch.run @@ fun sw ->
  let flow = Net.connect ~sw net addr in
  Flow.copy_string "GET / HTTP/1.0\r\n\r\n" flow;
  Flow.shutdown flow `Send;
  Flow.copy flow stdout

exception Connection_reset of exn

module Ipaddr : sig ... end

IP addresses.

module Sockaddr : sig ... end

Network addresses.

Provider Interfaces

class virtual listening_socket : object ... end

class virtual t : object ... end

Out-bound Connections

val connect : sw:Switch.t -> t -> Sockaddr.t -> < Flow.two_way; Flow.close; >

connect ~sw t addr is a new socket connected to remote address addr.

The new socket will be closed when sw finishes, unless closed manually first.

Incoming Connections

val listen : ?reuse_addr:bool -> ?reuse_port:bool -> backlog:int -> sw:Switch.t -> t -> Sockaddr.t -> listening_socket

listen ~sw ~backlog t addr is a new listening socket bound to local address addr.

The new socket will be closed when sw finishes, unless closed manually first.

For (non-abstract) Unix domain sockets, the path will be removed afterwards.

parameter backlog
The number of pending connections that can be queued up (see listen(2)).

parameter reuse_addr
Set the Unix.SO_REUSEADDR socket option. For Unix paths, also remove any stale left-over socket.

parameter reuse_port
Set the Unix.SO_REUSEPORT socket option.

val accept : sw:Switch.t -> listening_socket -> < Flow.two_way; Flow.close; > * Sockaddr.t

accept ~sw socket waits until a new connection is ready on socket and returns it.

The new socket will be closed automatically when sw finishes, if not closed earlier. If you want to handle multiple connections, consider using accept_sub instead.

val accept_sub : sw:Switch.t -> listening_socket -> on_error:(exn -> unit) -> (sw:Switch.t -> < Flow.two_way; Flow.close; > -> Sockaddr.t -> unit) -> unit

accept socket fn accepts a connection and handles it in a new fibre.

After accepting a connection to socket, it runs fn ~sw flow client_addr in a new fibre, using Fibre.fork_on_accept.

flow will be closed automatically when the sub-switch is finished, if not already closed by then.