Module Eio_mock

Mocks for testing.

When testing an Eio program it is often convenient to use mock resources rather than real OS-provided ones. This allows precise control over the test, such as adding delays or simulated faults. You can always just implement the various Eio types directly, but this module provides some convenient pre-built mocks, and some helpers for creating your own mocks.

Mocks typically use Eio.traceln to record how they were used. This output can be recorded and compared against a known-good copy using e.g. ocaml-mdx.

Mocks may require configuration. For example, a source flow needs to know what data to return when the application reads from it. This can be done using the various on_* functions. For example:

let stdin = Eio_mock.Flow.make "stdin" in
let stdout = Eio_mock.Flow.make "stdout" in
Eio_mock.Flow.on_read stdin [
  `Return "chunk1";
  `Return "chunk2";
  `Raise End_of_file
];
Eio.Flow.copy stdin stdout

This will produce:

+stdin: read "chunk1"
+stdout: wrote "chunk1"
+stdin: read "chunk2"
+stdout: wrote "chunk2"

Configuration

module Action : sig ... end

Actions that can be performed by mock handlers.

module Handler : sig ... end

Control how a mock responds.

Pre-defined mocks

module Flow : sig ... end

Mock Eio.Flow sources and sinks.

module Net : sig ... end

Mock Eio.Net networks and sockets.

module Clock : sig ... end

A mock Eio.Time clock for testing timeouts.

module Domain_manager : sig ... end

A mock Eio.Domain_manager that runs everything in a single domain.

Backend for mocks

The mocks can be used with any backend, but if you don't need any IO then you can use this one to avoid a dependency on eio_main.

module Backend : sig ... end

A dummy Eio backend with no actual IO.

Mock errors

type Eio.Exn.Backend.t +=

| Simulated_failure
(*
A fake error code you can use for simulated faults.
*)