Module Eio_unix

Extension of Eio for integration with OCaml's Unix module.

Note that OCaml's Unix module is not safe, and therefore care must be taken when using these functions. For example, it is possible to leak file descriptors this way, or to use them after they've been closed, allowing one module to corrupt a file belonging to an unrelated module.

val await_readable : Unix.file_descr -> unit

await_readable fd blocks until fd is readable (or has an error).

val await_writable : Unix.file_descr -> unit

await_writable fd blocks until fd is writable (or has an error).

module FD : sig ... end

Convert between Unix.file_descr and Eio objects.

module Ipaddr : sig ... end

Convert between Eio.Net.Ipaddr and Unix.inet_addr.

val sleep : float -> unit

sleep d sleeps for d seconds, allowing other fibers to run. This is can be useful for debugging (e.g. to introduce delays to trigger a race condition) without having to plumb Eio.Stdenv.clock through your code. It can also be used in programs that don't care about tracking determinism.

val run_in_systhread : ( unit -> 'a ) -> 'a

run_in_systhread fn runs the function fn in a newly created system thread (a Thread.t). This allows blocking calls to be made non-blocking.

module Private : sig ... end

API for Eio backends only.

module Ctf : sig ... end