Module Eio_unix

Extension of Eio for integration with OCaml's Unix module.

Note that OCaml's Unix module is not safe, and therefore care must be taken when using these functions. For example, it is possible to leak file descriptors this way, or to use them after they've been closed, allowing one module to corrupt a file belonging to an unrelated module.

type Eio.Exn.Backend.t +=

| Unix_error of Unix.error * string * string
(*
Wrapper for embedding Unix.Unix_error errors.
*)

module Fd : sig ... end

A safe wrapper for Unix.file_descr.

module Resource : sig ... end

Eio resources backed by an OS file descriptor.

module Net : sig ... end

Extended network API with support for file descriptors.

type source_ty = [

| `Unix_fd

| Eio.Resource.close_ty

| Eio.Flow.source_ty

]

type sink_ty = [

| `Unix_fd

| Eio.Resource.close_ty

| Eio.Flow.sink_ty

]

type 'a source = [> source_ty ] as 'a Eio.Std.r

type 'a sink = [> sink_ty ] as 'a Eio.Std.r

val await_readable : Unix.file_descr -> unit

await_readable fd blocks until fd is readable (or has an error).

val await_writable : Unix.file_descr -> unit

await_writable fd blocks until fd is writable (or has an error).

val sleep : float -> unit

sleep d sleeps for d seconds, allowing other fibers to run. This is can be useful for debugging (e.g. to introduce delays to trigger a race condition) without having to plumb Eio.Stdenv.mono_clock through your code. It can also be used in programs that don't care about tracking determinism.

val run_in_systhread : (unit -> 'a) -> 'a

run_in_systhread fn runs the function fn in a newly created system thread (a Thread.t). This allows blocking calls to be made non-blocking.

val pipe : Eio.Std.Switch.t -> source_ty Eio.Std.r * sink_ty Eio.Std.r

pipe sw returns a connected pair of flows src and sink. Data written to sink can be read from src. Note that, like all FDs created by Eio, they are both marked as close-on-exec by default.

module Process : sig ... end

Spawning child processes with extra control.

module Stdenv : sig ... end

The set of resources provided to a process on a Unix-compatible system.

module Private : sig ... end

API for Eio backends only.

module Pi : sig ... end