Module Eio_unix.Stdenv

The set of resources provided to a process on a Unix-compatible system.

type base = < stdin : source_ty Eio.Std.r ; stdout : sink_ty Eio.Std.r ; stderr : sink_ty Eio.Std.r ; net : [ `Unix | `Generic ] Eio.Net.ty Eio.Std.r ; domain_mgr : Eio.Domain_manager.ty Eio.Std.r ; process_mgr : Process.mgr_ty Eio.Std.r ; clock : float Eio.Time.clock_ty Eio.Std.r ; mono_clock : Eio.Time.Mono.ty Eio.Std.r ; fs : Eio.Fs.dir_ty Eio.Path.t ; cwd : Eio.Fs.dir_ty Eio.Path.t ; secure_random : Eio.Flow.source_ty Eio.Std.r ; debug : Eio.Debug.t ; backend_id : string >

The common set of features provided by all traditional operating systems (BSDs, Linux, Mac, Windows).

You can use the functions in Eio.Stdenv to access these fields if you prefer.